change int to float in result.content

+++ getipintel.py
@@ -9,7 +9,7 @@ def checkIP(ip):
 	timeout=5.00
 	#if you wish to use flags or json format, edit the request below
 	result = requests.get("http://check.getipintel.net/check.php?ip="+ip+"&contact="+contactEmail, timeout=timeout)
-	if (result.status_code != 200) or (int(result.content) < 0):
+	if (result.status_code != 200) or (float(result.content) < 0):
 		sys.stderr.write("An error occured while querying GetIPIntel")
 	if (float(result.content) > maxProbability):
 		return 1;
